CASP-828 Fix master wss script (#323)

+++ ci/wss.sh
@@ -36,21 +36,29 @@ fi
 pip install -U pip virtualenv >& /dev/null
 
 export PRODUCT_NAME=snowflake-connector-python
-export PROJECT_NAME=snowflake-connector-python
 
-DATE=$(date +'%m-%d-%Y')
+export PROD_BRANCH=master
+export PROJECT_VERSION=$TRAVIS_COMMIT
+
+env | grep TRAVIS | sort
+
+if [[ "$TRAVIS_PULL_REQUEST" != "false" ]]; then
+    echo "[INFO] Pull Request"
+    export PROJECT_NAME=PR-$TRAVIS_PULL_REQUEST
+elif [[ "$TRAVIS_BRANCH" == "$PROD_BRANCH" ]]; then
+    echo "[INFO] Production branch"
+    export PROJECT_NAME=$PROD_BRANCH
+else
+    echo "[INFO] Non Production branch. Skipping wss..."
+    export PROJECT_NAME=
+fi
 
 SCAN_DIRECTORIES=$(cd $THIS_DIR/.. && pwd)
 
-rm -f wss-unified-agent.jar
-curl -LO https://github.com/whitesource/unified-agent-distribution/releases/latest/download/wss-unified-agent.jar
-
-# whitesource will scan the folder and detect the corresponding configuration
-# configuration file wss-generated-file.config will be generated under ${SCAN_DIRECTORIES}
-# java -jar wss-unified-agent.jar -detect -d ${SCAN_DIRECTORIES}
-# SCAN_CONFIG="${SCAN_DIRECTORIES}/wss-generated-file.config"
-
-# SCAN_CONFIG is the path to your whitesource configuration file
+if [[ -n "$PROJECT_NAME" ]]; then
+    rm -f wss-unified-agent.jar
+    curl -LO https://github.com/whitesource/unified-agent-distribution/releases/latest/download/wss-unified-agent.jar
+fi
 SCAN_CONFIG=wss-agent.config
 cat > $SCAN_CONFIG <<CONFIG
 ###############################################################
@@ -141,29 +149,59 @@ case.sensitive.glob=false
 followSymbolicLinks=true
 CONFIG
 
-echo "[INFO] Running wss.sh for ${PROJECT_NAME}-${PRODUCT_NAME} under ${SCAN_DIRECTORIES}"
-java -jar wss-unified-agent.jar -apiKey ${WHITESOURCE_API_KEY} \
-    -c ${SCAN_CONFIG} \
-    -project ${PROJECT_NAME} \
-    -product ${PRODUCT_NAME} \
-    -d ${SCAN_DIRECTORIES} \
-    -wss.url https://saas.whitesourcesoftware.com/agent \
-    -offline true
-
-if java -jar wss-unified-agent.jar -apiKey ${WHITESOURCE_API_KEY} \
-   -c ${SCAN_CONFIG} \
-   -project ${PROJECT_NAME} \
-   -product ${PRODUCT_NAME} \
-   -projectVersion baseline \
-   -requestFiles whitesource/update-request.txt \
-   -wss.url https://saas.whitesourcesoftware.com/agent ; then
-    echo "checkPolicies=false" >> ${SCAN_CONFIG}
+set +e
+echo "[INFO] Running wss.sh for ${PRODUCT_NAME}-${PROJECT_NAME} under ${SCAN_DIRECTORIES}"
+if [[ "$PROJECT_NAME" == "$PROD_BRANCH" ]]; then
+    # Prod branch
     java -jar wss-unified-agent.jar -apiKey ${WHITESOURCE_API_KEY} \
         -c ${SCAN_CONFIG} \
-        -project ${PROJECT_NAME} \
+        -d ${SCAN_DIRECTORIES} \
         -product ${PRODUCT_NAME} \
-        -projectVersion ${DATE} \
-        -requestFiles whitesource/update-request.txt \
-        -wss.url https://saas.whitesourcesoftware.com/agent
+        -project ${PROJECT_NAME} \
+        -projectVersion ${PROJECT_VERSION} \
+        -offline true
+    ERR=$?
+    if [[ "$ERR" != "254" && "$ERR" != "0" ]]; then
+        echo "failed to run wss for $PRODUCT_VERSION_${PROJECT_VERSION} in ${PROJECT_VERSION}..."
+        exit 1
+    fi
+
+    java -jar wss-unified-agent.jar -apiKey ${WHITESOURCE_API_KEY} \
+       -c ${SCAN_CONFIG} \
+       -product ${PRODUCT_NAME} \
+       -project ${PROJECT_NAME} \
+       -projectVersion baseline \
+       -requestFiles whitesource/update-request.txt
+    ERR=$?
+    if [[ "$ERR" != "254" && "$ERR" != "0" ]]; then
+        echo "failed to run wss for $PRODUCT_VERSION_${PROJECT_VERSION} in baseline"
+        exit 1
+    fi
+    java -jar wss-unified-agent.jar -apiKey ${WHITESOURCE_API_KEY} \
+        -c ${SCAN_CONFIG} \
+        -product ${PRODUCT_NAME} \
+        -project ${PROJECT_NAME} \
+        -projectVersion ${PROJECT_VERSION} \
+        -requestFiles whitesource/update-request.txt
+    ERR=$?
+    if [[ "$ERR" != "254" && "$ERR" != "0" ]]; then
+        echo "failed to run wss for $PRODUCT_VERSION_${PROJECT_VERSION} in ${PROJECT_VERSION}"
+        exit 1
+    fi
+elif [[ -n "$PROJECT_NAME" ]]; then
+    # PR
+    java -jar wss-unified-agent.jar -apiKey ${WHITESOURCE_API_KEY} \
+        -c ${SCAN_CONFIG} \
+        -d ${SCAN_DIRECTORIES} \
+        -product ${PRODUCT_NAME} \
+        -project ${PROJECT_NAME} \
+        -projectVersion ${PROJECT_VERSION}
+    ERR=$?
+    if [[ "$ERR" != "254" && "$ERR" != "0" ]]; then
+        echo "failed to run wss for $PRODUCT_VERSION_${PROJECT_VERSION} in ${PROJECT_VERSION}..."
+        exit 1
+    fi
 fi
+set -e
+
 [[ -n "$IS_IN_CUSTOM_VENV" ]] && deactivate || true
